hw06/src/main/java/ru/otus/hw/repositories/JpaGenreRepository.java
package ru.otus.hw.repositories;

import jakarta.persistence.EntityManager;
import jakarta.persistence.PersistenceContext;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Repository;
import org.springframework.transaction.annotation.Transactional;
import ru.otus.hw.models.Genre;

import java.util.List;
import java.util.Set;

@Repository
@RequiredArgsConstructor
public class JpaGenreRepository implements GenreRepository {

    @PersistenceContext
    private final EntityManager em;

    @Override
    @Transactional(readOnly = true)
    public List<Genre> findAll() {
        return em.createQuery("SELECT g FROM Genre g", Genre.class)
                .getResultList();
    }

    @Override
    @Transactional(readOnly = true)
    public List<Genre> findAllByIds(Set<Long> ids) {
        return em.createQuery("SELECT g FROM Genre g WHERE g.id IN (:ids)", Genre.class)
                .setParameter("ids", ids)
                .getResultList();

    }
}